backend_python/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
async-timeout==4.0.3
attrs==25.3.0
cachetools==6.0.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
cssselect==1.3.0
cssutils==2.11.1
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.1
ftfy==6.3.1
GPUtil
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.3
idna==3.10
imagesize==1.4.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-community==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.7.0
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.4.2
numpy
openai==1.63.2
opencv-contrib-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.18
packaging==24.2
paddleocr==3.0.0
paddlepaddle==3.0.0
paddlex==3.0.0
pandas==1.5.3
pillow==11.2.1
premailer==3.10.0
prettytable==3.16.0
propcache==0.3.1
protobuf==6.31.1
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
ruamel.yaml==0.18.12
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.3
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.19.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
yarl==1.20.0
googletrans-py
